sensorcalitool: optimize the data collection
(1) you can define APPLY_DATA_REASSIGN macro to change max absolute
value to 46000 in step 4 and step5;
(2) default(now) it check whether the max absolute value less than 45000,
    if less, return -1, and the apk will tell you to do this step again
(3) the apk only display the data that was readed in the step, so when the
calibrate step end, the data in apk will not update.
(4) after the end of each step , the apk will display the result in the apk.
	the 1th step:
		acclzx --> accx(the apk display the data here)
		acclzy --> accy
		acclzz --> accz
		gyrox
		gyroy
		gyroz
		magnxnx
		magnxny
		magnxnz
	the 2th step:
		magnxsx
		magnxsy
	the 3th step:
		magnxsz
	the 4th step:
		gyrozx
		gyrozy
		gyrozz
	the 5th step:
		gyroyx
		gyroyy
		gyroyz
		acclyx
                acclyy
                acclyz

Change-Id: Iad2c2df33bb72bca48027f495ab38b2d5de46f01
Tracked-On:

Conflicts:
	cmds/sensorcalitool/calibrate.cpp